requirements.txt
langchain>=0.1.0
langchain-groq>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.2
streamlit>=1.30.0
requests>=2.31.0
tenacity>=8.2.3
jsonschema>=4.20.0
python-dotenv>=1.0.0
pytest>=7.4.3
notebook>=7.0.6
pandas>=2.0.0
plotly>=5.18.0
